docvault/requirements/base.txt
Django>=5.2,<5.3
psycopg[binary]>=3.3,<4
django-environ>=0.12
gunicorn>=25.0
djangorestframework>=3.16
django-allauth>=65.0
django-guardian>=3.3
django-cors-headers>=4.9
whitenoise>=6.11
drf-spectacular>=0.28
django-filter>=24.0
celery>=5.6
redis>=5.2
channels>=4.2
channels-redis>=4.2
django-storages>=1.14
boto3>=1.40
python-magic-bin>=0.4
ocrmypdf>=16.0
pikepdf>=9.0
Pillow>=12.0
langdetect>=1.0
dateparser>=1.3
extract-msg>=0.55
beautifulsoup4>=4.12
pdf2image>=1.17
Jinja2>=3.1
django-mptt>=0.16
elasticsearch>=8.0
scikit-learn>=1.7
nltk>=3.9
zxing-cpp>=2.2
openai>=1.0
ollama>=0.4
faiss-cpu>=1.9
numpy>=2.0
tiktoken>=0.7
pycryptodome>=3.23
python-gnupg>=0.5
mozilla-django-oidc>=5.0
pyotp>=2.9
qrcode>=8.0
spacy>=3.7
scikit-image>=0.22
PyMuPDF>=1.24
reportlab>=4.0
